pyproject.toml
[project]
name = "fastapi-clean"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13.0",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "python-decouple>=3.8",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "scalar-fastapi>=1.2.3",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn>=0.35.0",
]
